requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langsmith>=0.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0 